Spring Security JWT REDIS

+++ src/main/java/DealerStat/service/CommentService.java
@@ -4,12 +4,14 @@ import DealerStat.dto.CommentDto;
 import DealerStat.dto.MyUserDto;
 import DealerStat.entity.Comment;
 import DealerStat.entity.MyUser;
+import DealerStat.entity.Role;
 import DealerStat.repository.CommentRepository;
 import DealerStat.repository.MyUserRepository;
+import DealerStat.security.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Service
@@ -22,19 +24,32 @@ public class CommentService {
 
     private final MyUserService myUserService;
 
-    public Comment createComment(CommentDto commentDto, Long traiderId) {
-        Comment comment = new Comment();
-        comment.setMessage(commentDto.getMessage());
-        comment.setTrader(myUserRepository.findMyUserById(traiderId));
-        return commentRepository.save(comment);
+    private final JwtTokenProvider jwtTokenProvider;
+
+
+
+    public Comment createComment(CommentDto commentDto, Long traderId, HttpServletRequest request) {
+        if (commentDto.getRating() <= 5.0 && commentDto.getRating() >= 0) {
+            Comment comment = new Comment();
+            comment.setMessage(commentDto.getMessage());
+            comment.setAuthor(myUserRepository.findMyUserById(jwtTokenProvider.getId(request)));
+            comment.setTrader(myUserRepository.findMyUserById(traderId));
+            comment.setRating(commentDto.getRating());
+            return commentRepository.save(comment);
+        } else {
+            throw new RuntimeException("Rating must be >0 and <5");
+        }
+
     }
 
-    public Comment createCommentAndTrader(CommentDto commentDto, MyUserDto myUserDto){
-        myUserService.createUser(myUserDto);
-        Comment comment = new Comment();
-        comment.setMessage(commentDto.getMessage());
-        comment.setTrader(myUserRepository.findMyUserByFirstName(myUserDto.getFirstName()));
-        return commentRepository.save(comment);
+    public Comment createCommentAndTrader(CommentDto commentDto, MyUserDto myUserDto, HttpServletRequest request) {
+        if (commentDto.getRating() <= 5.0 && commentDto.getRating() >= 0) {
+//            myUserService.registerUser(myUserDto);
+            MyUser myUser = myUserRepository.findMyUserByEmail(myUserDto.getEmail());
+            return createComment(commentDto, myUser.getId(), request);
+        } else {
+            throw new RuntimeException("Rating must be >0 and <5");
+        }
     }
 
     public Comment showComment(Long id) {
@@ -45,22 +60,28 @@ public class CommentService {
         return commentRepository.findAllByTraderIdAndIsApprovedIsTrue(id);
     }
 
-    public Comment approveComment(Long id) {
+
+    public Comment updateComment(CommentDto commentDto, Long id) {
+        if (commentDto.getRating() <= 5 && commentDto.getRating() >= 0) {
+            Comment comment = commentRepository.findCommentById(id);
+            comment.setMessage(commentDto.getMessage());
+            comment.setRating(commentDto.getRating());
+            commentRepository.save(comment);
+            myUserService.refreshRating(comment.getId());
+            return comment;
+        } else {
+            throw new RuntimeException("Rating should be >0 and <5");
+        }
+    }
+
+    public void deleteComment(Long id, HttpServletRequest request) {
         Comment comment = commentRepository.findCommentById(id);
-        comment.setApproved(true);
-        return commentRepository.save(comment);
+        if (jwtTokenProvider.getId(request).equals(comment.getAuthor().getId())
+                || (jwtTokenProvider.getRole(request).contains(Role.ADMIN))) {
+            commentRepository.deleteById(id);
+            myUserService.refreshRating(comment.getTrader().getId());
+        } else {
+            throw new RuntimeException("Access denied");
+        }
     }
-
-    public Comment updateComment(String message, Long id) {
-        Comment comment = commentRepository.findCommentById(id);
-        comment.setMessage(message);
-        comment.setUpdatedAt(LocalDateTime.now());
-        return commentRepository.save(comment);
-    }
-
-    public void deleteComment(Long id) {
-
-        commentRepository.deleteById(id);
-    }
-
 }
